Added a basic unit test for the gegl-buffer-double, the only concrete
Added a basic unit test for the gegl-buffer-double, the only concrete
class I have added at the moment.  Currently I used the horrible hack of
copying gegl-init, gegl-object and gegl-utils (and editing gegl-init) so
that I didn't have to link against libgegl.la.  Trying to link against
libgegl.la in the test program causes problems because GeglBuffer is in
both libgegl and libimage and I did't know how to make the test binary
use the version from libimage.  This problem will go away in the future
when the version of gegl-buffer.c in gegl/gegl is permantly substituted
with the version in gegl/gegl/image