Modifications to allow compiling with "-Wall -Werror" This is just a good
Modifications to allow compiling with "-Wall -Werror"  This is just a
good practice.  static functions that were defined but not used were
commented out and the comment //defined but not used was added to mark
the change.  The functions were not removed entirely because I didn't
have time to examine the files and determine if the functions _should_
be removed.