do not install the custom gui's for some apps by default. (point GEGL_PATH
* bin/node-editors/Makefile.am: do not install the custom gui's for
some apps by default. (point GEGL_PATH to this dir in addition to the
regular GEGL_PATH, or move the .so's to GEGL_PATH to enable them.) The
reason they are removed is that they result in unresolved symbols when
loaded by other GEGL using apps than the sample binary.

svn path=/trunk/; revision=1731