removed mutex. added asserts. indentation, added revision, reduced size of
* gegl/buffer/gegl-buffer.c: removed mutex.
* gegl/buffer/gegl-buffer-load.c: added asserts.
* gegl/buffer/gegl-buffer-index.h: indentation, added revision,
reduced size of tile entries to 64bytes. (strictly speaking, padding
should only be needed for the header with this format, since the sizes
of blocks on disk are specified (and you should now the sizes of the
structs you are reading into to avoid reading too much appended
properties you do not know how to handle.

svn path=/trunk/; revision=2209