fetch the 64x64 context buffer based on the specific requested pixel and
* gegl/buffer/gegl-sampler.c: (gegl_sampler_get_from_buffer): fetch
the 64x64 context buffer based on the specific requested pixel and
not the center of interpolation. We might want to split this down
so that we can ensure that the needed context exist in the 64x64
buffer, thus allowing the samplers to rely on the context being
available within a 64 pixels wide buffer.

svn path=/trunk/; revision=2487