gegl_path_parameter_get_names, made count argument be a pointer. Made an
* gegl/property-types/gegl-path.[ch]: (gegl_path_to_string),
(gegl_path_parse_string), (gegl_path_parameter_get_names):
gegl_path_parameter_get_names,
made count argument be a pointer. Made an initial implementation of
serialization for the path parameters. a '!' is followed by the name
of a parameter (up to the next space char). The path itself is then
encoded using '_' as the node instructions.

svn path=/trunk/; revision=2826