gegl: Don't assume cache is always used
In gegl_eval_visitor_visit_pad() we can not always assume that the
cache has been used as the output buffer, so move this logic to all
users of gegl_operation_context_get_target() instead, and only call
gegl_cache_computed() if it actually was the cache that _was_ the
output buffer.

This is admittely not a good looking way of doing this, but it is
another sign that the current caching strategy and maybe even graph
evaluation strategy might need to be redesigned.