Check cached parameters to avoid relinking nodes
When prepare is called, we have been relinking the
input/composite/output nodes dependent on whether there is an input.
This causes massive cache invalidations on every call to prepare,
halting forward progress of gegl_processor_render. Cache the previous
parameters and only relink the nodes if there has been a change.