Add gegl_extension_handler_register/get_saver()
Add gegl_extension_handler_register/get_saver(), which are analogous
to gegl_extension_handler_register/get() except they are for output
instead of input. Also add a warning when an extension handler is not
found.

Add a simple test for it and add a TODO about cleaning this up when
cleaning up gegl-plugin.h