Created dynamic save operation using save_handlers
Introduced gegl:save, an operation which dynamically instances the
correct operation to save to a given path using the save_handler
routines.

It is implemented in terms of a sink operation, rather than a
meta-operation, so that it works with existing code. In particular, the
processor code uses introspection, and processes sink differently to
other operations.