operations: Add gegl:rgbe-load and gegl:rgbe-save
Add a radiance (rgbe) HDR image format reader and writer. Implements
reading of uncompressed and new-rle scanlines, and writing of
uncompressed scanlines.

The RGBE load and save code is put in a separate library so it's easy
to move out of the GEGL tree or replace with something else.

Also add test cases for both loading and saving.