Added exposure combiner, and support tool, for HDR
The exp-combine operation takes n input pads named 'exposure-%u'. They
are used to estimate the camera response function, and rebuild a
representative high range image.

The exp_combine tool takes an output, and a series of input and EV
exposure pairs. It uses the exp-combine operator to produce a high
range output. This tool has a dependency to libexiv2, for which checks
have been added to configure.

A regression test is also included, wrapped in a shell script. If
libexiv2 is not present, the test will be skipped (and not fail).

NOTE: make check fails on some platforms, we need to fix this
somehow. Maybe add relative error tolerance in img_cmp?