operations: mantiuk06 tone mapping operator
Implements the paper 'A Perceptual Framework for Contrast Processing of
High Dynamic Range Images'. It is a global tone mapping operator, which
constrains the set of contrasts across the full range of spatial
frequencies.

The implementation is a reasonably straight forward port from pfstmo
based code. There has been no attempt to improve/clean-up the base
implementation in this iteration.