Remove gegl-simd.h
Replace the GCC vector extensions code with code that is friendlier to the
auto-vectorizer in gcc. This optimizes some areas that didn't have optimizations before as well as simplifies code elsewhere. The core ops that had SIMD optimizations are also triggering the auto-vectorizer now, performance difference
seems hard to measure.