Change default PNG compression level to 3
A compression level of 9 with large natural images such as
those GEGL processes consumes a *lot* of CPU time with
little benefit. Most of the wait time is then the PNG save time.
Those who want that extra mileage can specify the compression
level or use a utility such as pngcrush.