vignette: add new op for vignetting
A vignetting op that allows choosing between 3 vignette shapes, adjusting
softness gamma and the squeezing of the vignette op. With the current logic the
vignette still can go outside the canvas with radius=1.0 and some aspect
ratios, this should probably be considered a bug and the shape scaled down to
fit.