gegl_buffer_sample: change API
Made scale be a GeglMatrix2 instead of a gfloat, allowing to specify a more
correct sampling context. This is an API change, not many things are using the
GEGL so this should be fine. In GIMP there is one instance that needs updating
but it resides in #ifdef 0'd code so there should be no API conflict.