exr-save: add new op from Rasmus Hahn
Op from Rasmus Hahn - rassahah@googlemail.com, original email follows:

Hi,

i have written an operation for gegl that writes OpenEXR float images, since
i noticed there does not seem to be one already and i could use it. This
version is quite usable within it's limitations. Limitations currently are:

 - writes only 32 bit linear floating point images.
 - only writes Y or RGB images with optional alpha.
 - does not make use of OpenEXR specialties like layers, mipmaps
   and multichannel.

It does support the tiled format; for simplicity, however, tiles are always
square.

I would like to see it included in the distribution; for now i have attached
a patch, which creates a new file gegl/operations/external/exr-save.cc and
makes a modification to the Makefile.am to build it. The patch is against
version gegl-0.1.6, but works with the git-version (as of 2011-07-13), too.
Sorry, for not providing a URL, but i do not have one right now.

The built operator is named 'gegl:exr-save' and has two properties:
 path (string) - the filename to write to
 tile (int) - the width of a tile (used for both x and y size);
   the default is 0, which means a scanline file is written.

Any comments or suggestions for this?

Greetings - Rasmus