buffer: add gegl_buffer_set_format:
Set the babl format of the buffer, setting the babl format of the buffer
requires the new format to have exactly the same bytes per pixel as the
original format. If NULL is passed in the format of the buffer is reset to the
original format.

Returns: the new babl format or NULL if it the passed in buffer was
incompatible (then the original format is still used).