buffer: mark linear buffers to avoid COW
Doing COW on GeglBuffers backed by linear buffers will not work, since it
effectively might replace the initial allocated buffer with the new COW
copy - removing assumptions the user of the API might have that the passed
in buffer remains a valid way of accessing the data.

A proper linear backend with real flushing to the underlying memory could
fix this.