gegl-init: fix the "swapping to ram" fallback when GEGL_SWAP environment variable is set.
In such a case, the NULL `swap` variable was overwritten afterwards
even though it was already tested that the directory could not be
created; hence GEGL would soon crash with a IO error.
Makes the code consistently use gegl_swap_dir () which has the right
value.

Additionally when the swap directory cannot be created, free the string
before setting it to NULL.