operations: add exposure operation
This operation is usefull for manipulating an images contrast and brigness in a different way to the classical brightness/contrast. It hast three parameters, applied in the following order: a multiplier, working in a relative logarithmic way, an offset and a gamma correction. This is most important when working with high dynamic range images.

Referring to bug:
https://bugzilla.gnome.org/show_bug.cgi?id=690615