opencl: opencl kernels are now separated files
created a new folder in gegl to contain opencl kernels.
the point is to have kernels as .cl files that are
included (with #include).

also, added some perl scripts that take .cl sources and
generate C strings that are fed to the opencl compiler
in runtime.

Thanks RPG for help with perl.