Murder CFLAGS before gi-r-scanner sees them
It seems that when CFLAGS exists in the environment the
Makefile's version of it leaks into the shell introspection
is running in. This causes the string to get parsed twice
and confuses the escaped \s.

Really this is a horrid hack and someone who knows more
about autotools should find the correct way to fix it.