Add gegl_operation_meta_watch_node
Passing nodes to gegl_operation_meta_watch_node or
gegl_operation_meta_watch_nodes will cause them to
be removed when the operation is destroyed. This
prevents the meta operation's children from leaking
if the root node's operation changes.

The watch holds a weak reference to the watched node,
so it is also safe to remove them prior to destroying
the operation.