ink-simulator: add transmittances for cyan and yellow inks
The value used when spectra are computed like in the op; are the
transmittances; not the reflectances (reflectance = illuminant
* transmittance). The values for the yellow and cyan inks were
roughly copied from plots in a couple of papers; a proper tabular/numerical
source would be nice.

Also added a final parameter for each ink; a linear scaling amount; to scale
the influence of the ink channel.