ink-simulator: update engine of ink-simulator
The core spectral simulation has been split out into a tiny non-glib
dependant core; the header and cfile for this core is embedded inline
in the op source for now.

Printer specifications now have an additional inkN.black spectral/color
sample. This permits determining the scattering/opacity of the pigments in the
ink, it is currently used as a wavelength dependent opacity similar to the
previous; now deprecated o= global per ink opacity. Such on black substrate
measurements are also the data that would be needed to drive kubelka munk
based simulation model.