operations: Support CMYK in jpg-load
Before would fail with "attempted to load unsupported JPEG (components=4)"
Also changes the format selection logic to be based on the out_color_space
in JPEG, instead of deducing it from the number of channels.

Note: uses the 'naive' CMYK support in BABL, which looks to
match what other viewer programs like Firefox and Eye of Gnome does.