Patch to clear up defaults and description for mono-mixer UI
From the GIMP UI it's very tough to tell what the values represent, so changing 'Amount of [color]' to '[color] Channel Multiplier]' will help understanding of what the UI is representing.

To go with this, set default value to .333 so a default conversion is approximately luminosity neutral.

Lastly, pull the value range toward a reasonable range -10 to 10 is extreme