Add a sepia filter
This implements the filer of the same name in the Filter Effects
specification: http://dev.w3.org/fxtf/filters/

The document says that filter functions must operate in the sRGB color
space. However, it also describes the color-interpolation-filters
property, which is set to linear by default, for the filtering
primitives used to implement the functions. Since sepia is a thin
wrapper around the feColorMatrix primitive, we ignore what the
specification says about the filter function, and expose the
color-interpolation-filters property on the filter operation.

https://bugzilla.gnome.org/show_bug.cgi?id=750542