buffer: reintroduce zeroing of scratch buffer allocation
Without this, valgrind complains about accessing uninitialized memory, this is
pixels around the edges of the scratch buffer - for some babl conversions this
also leads to horizontal/vertical line trails around brushes moving over the
image. Further investigation into whether a manual clearing of this or
avodiding the computations at the edges are faster would be nice.