operations/transform/transform-core: Avoid conversions in fast paths
The fast paths either pass the input buffer on to the output (like
gegl:nop), or create a shifted output buffer from the input. Hence,
there is no need for any Babl format conversions.

We don't need to check for the identity matrix because it is also a
fast translate matrix. Add a comment to clarify that.

https://bugzilla.gnome.org/show_bug.cgi?id=758534