buffer: clear hot tile when removing a tile from the cache
When removing a tile from the cache in gegl_tile_handler_cache_void()
and gegl_tile_handler_cache_invalidate(), check if it's the
associated tile storage's hot tile, and clear the hot tile if it is.

Failing to do so keeps the hot tile pointing to an outdated tile,
which can cause the 1x1 buffer access functions (_get_pixel(),
_set_pixel(), NN sampling) to return wrong results.

See the next commit for a test case.