text: add properties to control vertical size and alignemnt
The "vertical_alignment" property controls whether the text is
top-, middle-, or bottom-aligned; the text is aligned against the
height specified by the "vertical_wrap" property.

This commit also modifies the reported bounding-box of the op
to match the actual bounding box of the text, in particular, by
not having its top-left corner fixed at (0, 0).