crop: clip crop rect to input extents
If we don't do that, then the operation's output buffer, which is
a subbuffer of the input buffer, may contain regions that are not
backed by any storage, when the crop rect is not a subset of the
input buffer's extents.  If the output buffer is then resued for
in-place output as part of a following node, on the premise that
it actually is as big as it pretends to be, then any output
written to these regions is erroneously discarded.