crop: don't depend on the input node's bounding box ...
... in get_invalidated_by_change() and get_required_for_output()

This partially reverts commit 09e6133a340b666245749b682d9b76dd8b0d26d7.

The dependency on the input's boudning box is mostly important for
calculating the size of the output buffer (see the commit message of
the aforementioned commit), and not strictly required for these
functions (altough possibly more accurate).  This dependency did,
however, cause at least one issue (see previous commit), and the
calculation of the input's bounding box can be more expensive than
we'd like here, so reverting it for now, to stay on the safe side.