operations/core/load: Detect local ARWs and CR2s as RAW, not TIFF
This changes the content type guessing for local paths to match the
logic in GLib so that our interpretation of the content matches with
those of other components. We need to first look at the filename and
sniff the content only if it is inconclusive.

The code path dealing with URIs wasn't touched because the impact
wasn't clear.

Fallout from 442f6832701c4fba63a3564e6dd5ff1b9583a3a6

https://bugzilla.gnome.org/show_bug.cgi?id=771961