operations/external/raw-load: Improve the fidelity of the decoding
The output of the RAW loader is visibly different from that of similar
applications like Darktable and Shotwell. The decoded pixels are also
substantially different from their camera generated JPEG counterparts.
Let's address this by using similar parameters as Shotwell to configure
LibRaw.

Unlike Shotwell, we set the gamma values to (1.0, 1.0) for a linear
curve since our output format is linear RGB. Shotwell's use of
libraw_output_params_t.use_camera_matrix turned out to be ambiguous,
so, for the time being, we have opted for the default value.

https://bugzilla.gnome.org/show_bug.cgi?id=778183