operations/external/raw-load: Switch the output format to "R'G'B' u16"
A non-destructive image editor would want to perform certain operations
as early as possible in the pipeline. One such operation is
gegl:noise-reduction, and that works in "R'G'B'A float". Thus, it
makes sense for the RAW loader's output to be in the sRGB colour space
to avoid a needless conversion from perceptual to linear spaces.

Secondly, other commonly used loaders, like JPEG and PNG, also output
sRGB. Having the RAW loader use the same colour space will make it easy
for an application by removing the need to special case the
configuration of its pipeline for different image formats.

https://bugzilla.gnome.org/show_bug.cgi?id=778183