operations/workshop/dt-exposure: Add a srgb flag, but default to linear
Darktable's exposure filter is colourspace-agnostic and works in
"whatever colourspace the camera is in". For RAW images it operates in
linear RGB; and for JPEGs in sRGB or any other gamma corrected space
that might be specified in the file. Hence the output of the filter is
visibly different depending on whether it is RAW or JPEG.

However, Darktable is optimized for RAW images, and its exposure
filter is designed accordingly. It's reference behaviour is with RAW,
and the developers don't care much about JPEGs.

Therefore, let's default to linear RGB and add an option to use sRGB
in case someone wants to reproduce Darktable's handling of JPEGs.