operations/common/exposure: Add a property for adjusting blacks
If we only consider the 'exposure' property, then this operation is the
same as dt:exposure. One desirable thing about Darktable's version is
that it let's us adjust the black level - the threshold at which dark
grey values are cut off to pure black [1].

Let's add a property to offer the same feature.

If we leave black at its default value of 0.0, the behaviour of the
operation should remain largely unchanged. The only difference is that
the value of 'gain' won't exceed 100, while in the past it used to be
unbounded. This seems like an acceptable compromise.

[1] https://www.darktable.org/usermanual/ch03s04.html.php

https://bugzilla.gnome.org/show_bug.cgi?id=778366