median-blur: various additions
Performance
-----------

General performance improvement.  The speedup factor depends on
the input image and the radius, but it ranges from about 5x for
small radii, to 2x for large radii.

The main boost comes from keeping track of the last computed
median, and starting the histogram scan for the new median from
it, instead of from 0.  This is a heuristic, based on the
assumption that the median value changes very little across
consecutive pixels.  This results in the median search
approaching an amortized complexity of O(1), or at least a very
small O(n) (where n is the number of bins), while keeping the
cost of adding and removing values to the histogram O(1).

There are a few other smaller optimizations, the most
significant of which is probably converting input values to
bin indices only once per pixel, instead of twice.

Alpha
-----

Properly blur images with transparent regions.

Properties
----------

Add a few properties:

  - Neighborhood:  Controls the shape of the neighborhood.
    Can be either SQUARE (old behavior), CIRCLE (new default),
    or DIAMOND.

  - Percentile:  Which percentile of the neighborhood to use as
    the result, for the color channels.  Defaults to 50% (i.e.,
    the median), but can be anything from 0% (erosion), to 100%
    (dilation).  This, ofcourse, makes the name of the operation
    a misnomer :P  It's mostly for experimentation for now.

  - Alpha percentile:  Same thing, but for the alpha channel.