buffer: gegl_buffer_set, behavior change to expect roi prescaled for level
Uses of gegl_buffer_set with level == 0 does not change behavior and this is
the majority of existing uses.  This makes implementation of operations using
gegl_buffer_get and gegl_buffer_set more symmetric - permitting to use the same
GeglRectangle for gegl_buffer_set and gegl_buffer_get, a temporary copy is
avoided on almost every level aware call site for gegl_buffer_set. Since the
mipmap API hasn't been fully working in prior releases - this change can go out
as not an API change - but an improvement in expected results for the same
arguments as before.