ops: add hashes of reference composition for some ops
This is start of work integrate testing with the automatic operation gallery,
making it less neccesary to deal with automake and GEGL XML for having
regression tests in place. Using nice and short MD5 hashes since we're using it
as a checksum not to cryptographicly verify code or something we need to trust.