operation: Add autocleanups for GeglOperation sub-classes
Since GeglOperation sub-classes are reference-managed by GeglNode, it
is unlikely that applications will need to use g_autoptr with them.
This is primarily meant to let applications use G_DECLARE_* to reduce
the GObject boilerplate when defining their own operations. Since
G_DECLARE_* also defines the cleanup function for the new type, it is
necessary for the parent to also define it, even if it won't be used.

https://bugzilla.gnome.org/show_bug.cgi?id=780083