operations: sprinkle some GUM over a bunch of ops
GUM is a small language for controlling some UI attributes of operation
properties dynamically, based on context.  It is currently used by GIMP,
as of commit bb2417c29943e68baecd87f6c8c7808ebc3d091e.  See the comment
at the top of app/widget/gimppropgui-eval.c, in the GIMP tree, for a
description of the language.

This commit adds ui_meta keys containing GUM expression for a bunch of
ops, to make their UI more context aware.  The list of ops is not
necessarily exhaustive -- there may be more ops that can be GUMified.

Note that the new ui_meta keys augment, rather than replace, the existing
static attributes of properties (in particular, their label and
description), so that they still retain a meaningful value for users that
don't support GUM.