displace: fix alpha/abyss behavior; add center{,-{x,y}} props; gummify
Fix the way the displacement buffers' alpha affects the displacement
amount (probably a hiccup during conversion of old GIMP code).

Use ABYSS_NONE for the displacement buffers, instead of using the same
policy as for the input buffer.  The two are unrelated, and other
policies can cause artifacts.

Add a center, center-x and center-y properties.  Center is a boolean
flag that, when set, causes the displacement to be centered around
the point specified by center-x and center-y.  In cartesian mode,
this amounts to offsetting the displacement buffers so that they are
centered around that point; in polar mode, we also use that point as
the coordinate-system origin.

Gummify amount-x and amount-y to show more relevant labels/
descriptions.