warp: process stroke incrementally; misc improvements
When processing a stroke, keep a record of the processed path, and
keep the processed buffer around.  When the path changes, check if
the previously processed path is an initial segment of the new path,
in which case, only process the new points in the path, working on
top of the existing buffer.  As a result, since the GIMP warp tool
only ever appends points to the stroke path, it never has to
reprocess the entire path during a stroke, as it did before.

Cache the processed buffer ourselves, and pass it directly as output,
avoiding copying.

Stamp the first point of the path, and stroke segments such that the
final endpoint of the segment is always stamped, making the behvaior
more consistent.

Add a "spacing" property, controlling the stroke spacing, as a
fraction of the "brush" size.  This property defaults to 0.01, as
was previously hardcoded.  This allows GIMP to use higher values
for this property; 0.01 is too low for interactive use.